Add gerrit+spr+stgit flow commit hooks
If you have the gerrit commit hook installed, then "git spr" won't
work.  That is because the gerrit hook reformats "commit-id:UUID" to
"commit-id: UUID".

This adds a set of hook files to palves-hooks/ that can you can drop
in place of the gerrit hook to fix it.

Also, if you use stgit, then git spr's auto-rebasing breaks the stgit
flow.  Thankfully you can disable git spr auto-rebasing.  But if you
do that, then nothing adds the commit-id trailer to your commits, so
"git spr" won't work.  That is fixed by one of the hooks added here
taking care of adding the commit-id for you.  So when you edit or
create patches with stgit, the commit hooks are run automatically, and
you get the git spr trailer, just like you get a gerrit trailer.