Add a "[<series name> N/M] " prefix to the title of each PR in the stack
The "<series name>" part is taken from the first line of "git branch
--edit-description".

"git branch --edit-description" lets you edit the description
associated with the current branch (stored in .git/refs/heads/<branch>
metadata, under .git/refs/meta in config).  That description is what
"git format-patch --cover" is based on, so I think it's the fit place
to retrieve it from.

To show the description, you don’t use a git branch subcommand.
Instead, Git exposes this via:

 git config branch.<branch>.description

This is being done in UpdatePullRequest, which is called both when we
create the PRs the first time, and whenever there's a subsequent
update.  An improvement would be do to it in CreatePullRequest too,
because as is currently, if "git spr update" creates a PR, it first
creates it with title "foo", and then updates it to "[<series name>
N/M]" immediately after, and that is visible in the PR conversation
list in the web interface.

Should also probably move this new code to helpers.go.